conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
astroid=2.4.2=py37_0
blas=1.0=mkl
bzip2=1.0.8=h516909a_3
ca-certificates=2021.5.30=ha878542_0
cairo=1.16.0=hf32fb01_1
certifi=2021.5.30=py37h89c1867_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.10.0=py37_0
dbus=1.13.18=hb2f20db_0
expat=2.2.10=he6710b0_2
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.1=h6c09931_0
freeglut=3.2.1=h58526e2_0
freetype=2.10.4=h5ab3b9f_0
glib=2.68.2=h36276a3_0
graphite2=1.3.13=h58526e2_1001
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hc401514_3
icu=58.2=he6710b0_3
install=1.3.4=pypi_0
intel-openmp=2020.2=254
isort=5.6.4=py_0
jasper=2.0.14=h07fcdf6_1
joblib=1.0.0=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
kiwisolver=1.3.0=py37h2531618_0
lazy-object-proxy=1.4.3=py37h27cfd23_2
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran=3.0.0=1
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=he1b5a44_1001
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.2=heb0550a_3
matplotlib=3.3.2=h06a4308_0
matplotlib-base=3.3.2=py37h817c723_0
mccabe=0.6.1=py37_1
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
ncurses=6.2=he6710b0_1
ninja=1.10.2=py37hff7bd54_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
opencv-python=4.5.1.48=pypi_0
openssl=1.1.1k=h27cfd23_0
pandas=1.2.0=py37ha9443f7_0
pcre=8.44=he6710b0_0
pillow=8.1.0=py37he98fc37_0
pip=20.3.3=py37h06a4308_0
pixman=0.40.0=h36c2ea0_0
py-opencv=3.4.2=py37hb342d67_1
pylint=2.6.0=py37_0
pyparsing=2.4.7=py_0
pyqt=5.9.2=py37h05f1152_2
python=3.7.9=h7579374_0
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.7.1=py3.7_cuda11.0.221_cudnn8.0.5_0
pytz=2020.5=pyhd3eb1b0_0
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
scikit-learn=0.23.2=py37h0573a6f_0
scipy=1.5.2=py37h0b6359f_0
seaborn=0.11.1=pyhd3eb1b0_0
setuptools=51.1.2=py37h06a4308_4
sip=4.19.8=py37hf484d3e_0
six=1.15.0=py37h06a4308_0
sqlite=3.33.0=h62c20be_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
toml=0.10.1=py_0
torchaudio=0.7.2=py37
torchvision=0.8.2=py37_cu110
tornado=6.1=py37h27cfd23_0
tqdm=4.55.1=pyhd3eb1b0_0
typed-ast=1.4.2=py37h27cfd23_1
typing_extensions=3.7.4.3=py_0
wheel=0.36.2=pyhd3eb1b0_0
wrapt=1.11.2=py37h7b6447c_0
xorg-fixesproto=5.0=h14c3975_1002
xorg-inputproto=2.3.2=h14c3975_1002
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libx11=1.6.12=h36c2ea0_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxfixes=5.0.3=h516909a_1004
xorg-libxi=1.7.10=h516909a_0
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
